# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegra",
    "description": "Production-ready Agent Protocol server",
    "version": "0.17.0"
  },
  "paths": {
...
    "/runs/crons": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Create Cron",
        "description": "Create a cron job that fires on a schedule (stateless).\n\nPersists the cron record, then triggers the first run immediately and\nreturns the ``Run`` object (matching LangGraph SDK ``create()`` contract).\nWhen the caller passes ``enabled=False`` the first run is suppressed and\nthe response is the persisted ``Cron`` instead.",
        "operationId": "create_cron_runs_crons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Run"
                    },
                    {
                      "$ref": "#/components/schemas/CronResponse"
                    }
                  ],
                  "title": "Response Create Cron Runs Crons Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/runs/crons": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Create Cron For Thread",
        "description": "Create a cron job bound to an existing thread.\n\nThe thread is reused for every scheduled run. Triggers the first run\nimmediately and returns the ``Run`` object. When ``enabled=False`` is\npassed the first run is suppressed and the persisted cron is returned.",
        "operationId": "create_cron_for_thread_threads__thread_id__runs_crons_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Run"
                    },
                    {
                      "$ref": "#/components/schemas/CronResponse"
                    }
                  ],
                  "title": "Response Create Cron For Thread Threads  Thread Id  Runs Crons Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/{cron_id}": {
      "patch": {
        "tags": [
          "Crons"
        ],
        "summary": "Update Cron",
        "description": "Update an existing cron job.\n\nOnly provided fields are updated (partial patch). Returns the full\n``Cron`` object after update.",
        "operationId": "update_cron_runs_crons__cron_id__patch",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cron Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Crons"
        ],
        "summary": "Delete Cron",
        "description": "Delete a cron job.",
        "operationId": "delete_cron_runs_crons__cron_id__delete",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cron Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/search": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Search Crons",
        "description": "Search cron jobs with filters and pagination.\n\nA full page sets the `X-Pagination-Next` cursor header. Pass `select` to\nreturn only the listed fields.\n\nDeclared without a response model because `select` makes the row shape\ndynamic \u2014 full entities when omitted, projected dicts when given.",
        "operationId": "search_crons_runs_crons_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/count": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Count Crons",
        "description": "Count cron jobs matching filters.",
        "operationId": "count_crons_runs_crons_count_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronCountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Count Crons Runs Crons Count Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentProtocolError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "AgentProtocolError",
        "description": "Standard Agent Protocol error response"
      },
...
      "CronCountRequest": {
        "properties": {
          "assistant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistant Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Metadata containment match."
          }
        },
        "type": "object",
        "title": "CronCountRequest",
        "description": "Request body for counting cron jobs."
      },
      "CronCreate": {
        "properties": {
          "assistant_id": {
            "type": "string",
            "maxLength": 256,
            "title": "Assistant Id"
          },
          "schedule": {
            "type": "string",
            "maxLength": 256,
            "title": "Schedule"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "interrupt_before": {
            "anyOf": [
              {
                "type": "string",
                "const": "*"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Before"
          },
          "interrupt_after": {
            "anyOf": [
              {
                "type": "string",
                "const": "*"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt After"
          },
          "webhook": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook"
          },
          "on_run_completed": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "delete",
                  "keep"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "On Run Completed"
          },
          "multitask_strategy": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Multitask Strategy"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "stream_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Mode"
          },
          "stream_subgraphs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Subgraphs"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "stream_resumable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Resumable"
          },
          "durability": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sync",
                  "async",
                  "exit"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Durability"
          },
          "checkpoint_during": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint During",
            "description": "DEPRECATED: use `durability`.",
            "deprecated": true
          }
        },
        "type": "object",
        "required": [
          "assistant_id",
          "schedule"
        ],
        "title": "CronCreate",
        "description": "Request body for creating a cron job (stateless or thread-bound)."
      },
      "CronResponse": {
        "properties": {
          "cron_id": {
            "type": "string",
            "title": "Cron Id"
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "on_run_completed": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "delete",
                  "keep"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "On Run Completed"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "schedule": {
            "type": "string",
            "title": "Schedule"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "next_run_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run Date"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "cron_id",
          "assistant_id",
          "schedule",
          "created_at",
          "updated_at"
        ],
        "title": "CronResponse",
        "description": "Response model matching the SDK ``Cron`` TypedDict."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Run": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique identifier for the run."
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "Thread this run belongs to."
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id",
            "description": "Assistant that is executing this run."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current run status: pending, running, error, success, timeout, or interrupted.",
            "default": "pending"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input",
            "description": "Input data provided to the run. None for checkpoint-only resume."
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output",
            "description": "Final output produced by the run, or null if not yet complete."
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Error message if the run failed."
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Configuration passed to the graph at runtime."
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Context variables available during execution."
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Identifier of the user who owns this run."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the run was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Timestamp when the run was last updated."
          },
          "metadata_dict": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Dict",
            "description": "Metadata supplied when the run was created."
          },
          "multitask_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multitask Strategy",
            "description": "Strategy applied for concurrent runs on the same thread."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "thread_id",
          "assistant_id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "Run",
        "description": "Run entity model\n\nStatus values: pending, running, error, success, timeout, interrupted"
      },
...
    }
  },
...
}